+++ src/Framework.SecuritySystem/SecurityPath/Builders/V1_MaterializedPermissions/SecurityExpressionBuilderBase.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
@@ -9,7 +8,9 @@ using Framework.Core;
 using Framework.HierarchicalExpand;
 using Framework.Persistent;
 
-namespace Framework.SecuritySystem.Rules.Builders.V2
+using JetBrains.Annotations;
+
+namespace Framework.SecuritySystem.Rules.Builders.MaterializedPermissions
 {
     public abstract class SecurityExpressionBuilderBase<TPersistentDomainObjectBase, TDomainObject, TIdent>
         : ISecurityExpressionBuilder<TPersistentDomainObjectBase, TDomainObject, TIdent>
@@ -21,7 +22,7 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
         internal readonly SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> Factory;
 
         protected SecurityExpressionBuilderBase(
-            [JetBrains.Annotations.NotNull] SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory)
+            [NotNull] SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory)
         {
             this.Factory = factory ?? throw new ArgumentNullException(nameof(factory));
         }
@@ -33,20 +34,20 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
         }
 
 
-        public Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression<TSecurityOperation>(ContextSecurityOperation<TSecurityOperation> securityOperation)
-                where TSecurityOperation : struct, Enum
-        {
-            var filterExpression = this.GetSecurityFilterExpression(securityOperation.SecurityExpandType).ExpandConst().InlineEval();
-
-            var baseQuery = this.Factory.AuthorizationSystem.GetPermissionQuery(securityOperation);
-
-            return domainObject => baseQuery.Any(permission => filterExpression.Eval(domainObject, permission));
-        }
 
 
-        public abstract Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType);
+        public abstract Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(Dictionary<Type, IEnumerable<TIdent>> permission);
 
         public abstract Expression<Func<IPermission<TIdent>, bool>> GetAccessorsFilter(TDomainObject domainObject, HierarchicalExpandType expandType);
+
+        public abstract IEnumerable<Type> GetUsedTypes();
+
+        public virtual Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(List<Dictionary<Type, IEnumerable<TIdent>>> permissions)
+        {
+            if (permissions == null) throw new ArgumentNullException(nameof(permissions));
+
+            return permissions.BuildOr(this.GetSecurityFilterExpression);
+        }
 
         public Expression<Func<IEnumerable<IPermission<TIdent>>, bool>> GetAccessorsFilterMany(TDomainObject domainObject, HierarchicalExpandType expandType)
         {
@@ -65,16 +66,23 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
     {
         protected readonly TPath Path;
 
-        protected SecurityExpressionBuilderBase([JetBrains.Annotations.NotNull] SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory,
+        protected SecurityExpressionBuilderBase([NotNull] SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory,
                                                 TPath path) : base(factory)
         {
             this.Path = path ?? throw new ArgumentNullException(nameof(path));
         }
 
+
+        public override IEnumerable<Type> GetUsedTypes()
+        {
+            return this.Path.GetUsedTypes();
+        }
+
+
         public abstract class SecurityPathExpressionBuilderBase<TInnerPath> : SecurityExpressionBuilderBase<TPersistentDomainObjectBase, TDomainObject, TIdent, TInnerPath>
             where TInnerPath : SecurityPath<TPersistentDomainObjectBase, TDomainObject, TIdent>
         {
-            protected SecurityPathExpressionBuilderBase(SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory, TInnerPath path) : base(factory, path)
+            protected SecurityPathExpressionBuilderBase(SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory, TInnerPath path): base(factory, path)
             {
 
             }
@@ -88,6 +96,23 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 : base(factory, path)
             {
 
+            }
+
+
+            protected abstract Func<IEnumerable<TIdent>, Expression<Func<TDomainObject, bool>>> SecurityFilter { get; }
+
+
+
+            public sealed override Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(Dictionary<Type, IEnumerable<TIdent>> permission)
+            {
+                if (permission.TryGetValue(typeof(TSecurityContext), out var securityIdents))
+                {
+                    return this.SecurityFilter(securityIdents);
+                }
+                else
+                {
+                    return _ => true;
+                }
             }
 
             protected abstract IEnumerable<TSecurityContext> GetSecurityObjects(TDomainObject domainObject);
@@ -126,15 +151,17 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             public SecurityByIdentsExpressionBuilder(SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory, SecurityPath<TPersistentDomainObjectBase, TDomainObject, TIdent>.SecurityPathByIdents<TSecurityContext> path)
                 : base(factory, path)
             {
+
+            }
+
+
+            protected override Func<IEnumerable<TIdent>, Expression<Func<TDomainObject, bool>>> SecurityFilter
+            {
+                get { return this.Path.SecurityFilter; }
             }
 
 
             protected override IEnumerable<TSecurityContext> GetSecurityObjects(TDomainObject domainObject)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
             {
                 throw new NotImplementedException();
             }
@@ -145,13 +172,13 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             public ConditionBinarySecurityPathExpressionBuilder(SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory, SecurityPath<TPersistentDomainObjectBase, TDomainObject, TIdent>.ConditionPath path)
                 : base(factory, path)
             {
+
             }
 
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
-            {
-                var securityFilter = this.Path.SecurityFilter;
 
-                return (domainObject, _) => securityFilter.Eval(domainObject);
+            public override Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(Dictionary<Type, IEnumerable<TIdent>> _)
+            {
+                return this.Path.SecurityFilter;
             }
 
             public override Expression<Func<IPermission<TIdent>, bool>> GetAccessorsFilter(TDomainObject domainObject, HierarchicalExpandType expandType)
@@ -161,7 +188,7 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 return _ => hasAccess;
             }
 
-            [SuppressMessage("SonarQube", "S2743")]
+
             private static readonly LambdaCompileCache LambdaCompileCache = new LambdaCompileCache(LambdaCompileMode.All);
         }
 
@@ -174,50 +201,26 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             }
 
 
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
+            protected override Func<IEnumerable<TIdent>, Expression<Func<TDomainObject, bool>>> SecurityFilter
             {
-                var entityTypeId = this.Factory.AuthorizationSystem.ResolveSecurityTypeId(typeof(TSecurityContext));
-
-                var eqIdentsExpr = ExpressionHelper.GetEquality<TIdent>();
-
-                var getIdents = ExpressionHelper.Create((IPermission<TIdent> permission) =>
-                    permission.FilterItems
-                              .Select(fi => fi.Entity)
-                              .Where(item => eqIdentsExpr.Eval(item.EntityType.Id, entityTypeId))
-                              .Select(pfe => pfe.EntityId))
-                                                .ExpandConst()
-                                                .InlineEval();
-
-                var expander = this.Factory.HierarchicalObjectExpanderFactory.CreateQuery(typeof(TSecurityContext));
-
-                var expandExpression = expander.GetExpandExpression(expandType);
-
-                var expandExpressionQ = from idents in getIdents
-                                        select expandExpression.Eval(idents);
-
-                switch (this.Path.Mode)
+                get
                 {
-                    case SingleSecurityMode.AllowNull:
+                    switch (this.Path.Mode)
+                    {
+                        case SingleSecurityMode.AllowNull:
 
-                        return (domainObject, permission) =>
+                            return securityIdents => from securityObject in this.Path.SecurityPath
 
-                                       !getIdents.Eval(permission).Any()
+                                                     select securityObject == null || securityIdents.Contains(securityObject.Id);
 
-                                    || this.Path.SecurityPath.Eval(domainObject) == null
+                        case SingleSecurityMode.Strictly:
 
-                                    || expandExpressionQ.Eval(permission).Contains(this.Path.SecurityPath.Eval(domainObject).Id);
+                            return securityIdents => from securityObject in this.Path.SecurityPath
 
-                    case SingleSecurityMode.Strictly:
+                                                     select securityIdents.Contains(securityObject.Id);
 
-                        return (domainObject, permission) =>
-
-                                       !getIdents.Eval(permission).Any()
-
-                                    || expandExpressionQ.Eval(permission).Contains(this.Path.SecurityPath.Eval(domainObject).Id);
-
-                    default:
-
-                        throw new ArgumentOutOfRangeException(this.Path.Mode.ToString());
+                        default: throw new ArgumentOutOfRangeException(this.Path.Mode.ToString());
+                    }
                 }
             }
 
@@ -233,7 +236,6 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 }
             }
 
-            [SuppressMessage("SonarQube", "S2743")]
             private static readonly LambdaCompileCache LambdaCompileCache = new LambdaCompileCache(LambdaCompileMode.All);
         }
 
@@ -244,48 +246,29 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             public ManySecurityExpressionBuilder(SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> factory, SecurityPath<TPersistentDomainObjectBase, TDomainObject, TIdent>.ManySecurityPath<TSecurityContext> path)
                 : base(factory, path)
             {
+
             }
 
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
+
+            protected override Func<IEnumerable<TIdent>, Expression<Func<TDomainObject, bool>>> SecurityFilter
             {
-                var entityTypeId = this.Factory.AuthorizationSystem.ResolveSecurityTypeId(typeof(TSecurityContext));
-
-                var eqIdentsExpr = ExpressionHelper.GetEquality<TIdent>();
-
-                var getIdents = ExpressionHelper.Create((IPermission<TIdent> permission) =>
-                                                                permission.FilterItems
-                                                                          .Select(fi => fi.Entity)
-                                                                          .Where(item => eqIdentsExpr.Eval(item.EntityType.Id, entityTypeId))
-                                                                          .Select(pfe => pfe.EntityId))
-                                                .ExpandConst()
-                                                .InlineEval();
-
-                var expander = (IHierarchicalObjectQueryableExpander<TIdent>)this.Factory.HierarchicalObjectExpanderFactory.Create(typeof(TSecurityContext));
-
-                var expandExpression = expander.GetExpandExpression(expandType);
-
-                var expandExpressionQ = from idents in getIdents
-                                        select expandExpression.Eval(idents);
-
-                switch (this.Path.Mode)
+                get
                 {
-                    case ManySecurityPathMode.AnyStrictly:
+                    switch (this.Path.Mode)
+                    {
+                        case ManySecurityPathMode.AnyStrictly:
                         {
                             if (this.Path.SecurityPathQ != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPathQ
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || this.Path.SecurityPathQ.Eval(domainObject).Any(item => expandExpressionQ.Eval(permission).Contains(item.Id));
+                                                         select securityObjects.Any(item => securityIdents.Contains(item.Id));
                             }
                             else if (this.Path.SecurityPath != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPath
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || this.Path.SecurityPath.Eval(domainObject).Any(item => expandExpressionQ.Eval(permission).Contains(item.Id));
+                                                         select securityObjects.Any(item => securityIdents.Contains(item.Id));
                             }
                             else
                             {
@@ -293,27 +276,19 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                             }
                         }
 
-                    case ManySecurityPathMode.Any:
+                        case ManySecurityPathMode.Any:
                         {
                             if (this.Path.SecurityPathQ != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPathQ
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || !this.Path.SecurityPathQ.Eval(domainObject).Any()
-
-                                               || this.Path.SecurityPathQ.Eval(domainObject).Any(item => getIdents.Eval(permission).Contains(item.Id));
+                                                         select !securityObjects.Any() || securityObjects.Any(item => securityIdents.Contains(item.Id));
                             }
                             else if (this.Path.SecurityPath != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPath
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || !this.Path.SecurityPath.Eval(domainObject).Any()
-
-                                               || this.Path.SecurityPath.Eval(domainObject).Any(item => getIdents.Eval(permission).Contains(item.Id));
+                                                         select !securityObjects.Any() || securityObjects.Any(item => securityIdents.Contains(item.Id));
                             }
                             else
                             {
@@ -321,23 +296,19 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                             }
                         }
 
-                    case ManySecurityPathMode.All:
+                        case ManySecurityPathMode.All:
                         {
                             if (this.Path.SecurityPathQ != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPathQ
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || this.Path.SecurityPathQ.Eval(domainObject).All(item => getIdents.Eval(permission).Contains(item.Id));
+                                                         select securityObjects.All(item => securityIdents.Contains(item.Id));
                             }
                             else if (this.Path.SecurityPath != null)
                             {
-                                return (domainObject, permission) =>
+                                return securityIdents => from securityObjects in this.Path.SecurityPath
 
-                                               !getIdents.Eval(permission).Any()
-
-                                               || this.Path.SecurityPath.Eval(domainObject).All(item => getIdents.Eval(permission).Contains(item.Id));
+                                                         select securityObjects.All(item => securityIdents.Contains(item.Id));
                             }
                             else
                             {
@@ -345,9 +316,10 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                             }
                         }
 
-                    default:
+                        default:
 
-                        throw new ArgumentOutOfRangeException("this.Path.Mode");
+                            throw new ArgumentOutOfRangeException("this.Path.Mode");
+                    }
                 }
             }
 
@@ -356,7 +328,6 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 return this.Path.SecurityPath.Eval(domainObject, LambdaCompileCache).EmptyIfNull();
             }
 
-            [SuppressMessage("SonarQube", "S2743")]
             private static readonly LambdaCompileCache LambdaCompileCache = new LambdaCompileCache(LambdaCompileMode.All);
         }
 
@@ -365,12 +336,15 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             where TNestedObject : class, TPersistentDomainObjectBase
         {
             private readonly SecurityExpressionBuilderBase<TPersistentDomainObjectBase, TNestedObject, TIdent> _nestedBuilder;
-            
+
+            private static readonly MethodInfo anyEmptyEnumerableMethodInfo;
+            private static readonly MethodInfo anyEnumerableMethodInfo;
+            private static readonly MethodInfo allEnumerableMethodInfo;
             private static readonly string getAccessortFilterMethodInfoName;
+            private static readonly MethodInfo selectPermissionFuncEnumerableMethodInfo;
             private static readonly MethodInfo buildOrMethod;
             private static readonly MethodInfo buildAndMethod;
 
-            [SuppressMessage("SonarQube", "S2743")]
             private static readonly LambdaCompileCache LambdaCompileCache = new LambdaCompileCache(LambdaCompileMode.All);
 
             private readonly Lazy<Expression<Func<TDomainObject, HierarchicalExpandType, Expression<Func<IPermission<TIdent>, bool>>>>> _getAccessableFilterLazy;
@@ -379,9 +353,20 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
 
             static NestedManySecurityExpressionBuilder()
             {
+                anyEmptyEnumerableMethodInfo = new Func<IEnumerable<TNestedObject>, bool>(Enumerable.Any).Method;
+
+                anyEnumerableMethodInfo = new Func<IEnumerable<TNestedObject>, Func<TNestedObject, bool>, bool>(Enumerable.Any).Method;
+
+                allEnumerableMethodInfo = new Func<IEnumerable<TNestedObject>, Func<TNestedObject, bool>, bool>(Enumerable.All).Method;
+
                 buildOrMethod = ((Func<IEnumerable<Expression<Func<IPermission<TIdent>, bool>>>, Expression<Func<IPermission<TIdent>, bool>>>)(Framework.Core.ExpressionExtensions.BuildOr)).Method;
                 buildAndMethod = ((Func<IEnumerable<Expression<Func<IPermission<TIdent>, bool>>>, Expression<Func<IPermission<TIdent>, bool>>>)(Framework.Core.ExpressionExtensions.BuildAnd)).Method;
-                
+
+                selectPermissionFuncEnumerableMethodInfo = ((Func<
+                    IEnumerable<TNestedObject>,
+                    Func<TNestedObject, Expression<Func<IPermission<TIdent>, bool>>>,
+                    IEnumerable<Expression<Func<IPermission<TIdent>, bool>>>>)(Enumerable.Select)).Method;
+
                 getAccessortFilterMethodInfoName = "GetAccessorsFilter";
             }
 
@@ -394,30 +379,49 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 this._getAccessableFilterLazy = new Lazy<Expression<Func<TDomainObject, HierarchicalExpandType, Expression<Func<IPermission<TIdent>, bool>>>>>(() => this.CreateAccessorsFilterExpression(), true);
             }
 
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
+            public override Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(List<Dictionary<Type, IEnumerable<TIdent>>> permissions)
             {
-                var baseFilter = this._nestedBuilder.GetSecurityFilterExpression(expandType).ExpandConst().InlineEval();
-                
+                var filterExpression = permissions.BuildOr(this._nestedBuilder.GetSecurityFilterExpression);
+
                 switch (this.Path.Mode)
                 {
                     case ManySecurityPathMode.Any:
 
-                        return (domainObject, permission) => !this.Path.NestedObjectsPath.Eval(domainObject).Any()
+                        var any = Expression.Call(anyEmptyEnumerableMethodInfo, this.Path.NestedObjectsPath.Body);
+                        var anyFilter = Expression.Call(anyEnumerableMethodInfo, this.Path.NestedObjectsPath.Body, filterExpression);
 
-                                                          || this.Path.NestedObjectsPath.Eval(domainObject).Any(nestedObject => baseFilter.Eval(nestedObject, permission));
+                        var notAny = Expression.Not(any);
+                        var finalOr = Expression.OrElse(notAny, anyFilter);
+
+                        var anyResult = Expression.Lambda<Func<TDomainObject, bool>>(finalOr, this.Path.NestedObjectsPath.Parameters.First());
+
+                        return anyResult;
 
                     case ManySecurityPathMode.AnyStrictly:
 
-                        return (domainObject, permission) => this.Path.NestedObjectsPath.Eval(domainObject).Any(nestedObject => baseFilter.Eval(nestedObject, permission));
+                        var onlyAnyFilter = Expression.Call(anyEnumerableMethodInfo, this.Path.NestedObjectsPath.Body, filterExpression);
+
+                        var onlyAnyResult = Expression.Lambda<Func<TDomainObject, bool>>(onlyAnyFilter, this.Path.NestedObjectsPath.Parameters.First());
+
+                        return onlyAnyResult;
 
                     case ManySecurityPathMode.All:
 
-                        return (domainObject, permission) => this.Path.NestedObjectsPath.Eval(domainObject).All(nestedObject => baseFilter.Eval(nestedObject, permission));
+                        var allFilter = Expression.Call(allEnumerableMethodInfo, this.Path.NestedObjectsPath.Body, filterExpression);
+
+                        var allResult = Expression.Lambda<Func<TDomainObject, bool>>(allFilter, this.Path.NestedObjectsPath.Parameters.First());
+
+                        return allResult;
 
                     default:
 
                         throw new ArgumentOutOfRangeException("this.Path.Mode");
                 }
+            }
+
+            public override Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(Dictionary<Type, IEnumerable<TIdent>> permission)
+            {
+                throw new NotImplementedException();
             }
 
             public override Expression<Func<IPermission<TIdent>, bool>> GetAccessorsFilter(TDomainObject domainObject, HierarchicalExpandType expandType)
@@ -443,7 +447,7 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                 var getAccessorFilters = Expression.Call(
                     builderParameter,
                     getAccessortFilterMethodInfoName,
-                    Type.EmptyTypes,
+                    new Type[0],
                     nestedObjectParameter,
                     expandTypeParameter);
 
@@ -459,7 +463,7 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
                     this.Path.NestedObjectsPath.Body,
                     getAccessorFiltersExpression);
 
-                MethodInfo buildMethodInfo;
+                MethodInfo buildMethodInfo = null;
 
                 switch (this.Path.Mode)
                 {
@@ -502,6 +506,15 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             protected abstract Expression<Func<TArg, bool>> BuildOperation<TArg>(Expression<Func<TArg, bool>> arg1, Expression<Func<TArg, bool>> arg2);
 
 
+            public override Expression<Func<TDomainObject, bool>> GetSecurityFilterExpression(Dictionary<Type, IEnumerable<TIdent>> permission)
+            {
+                if (permission == null) throw new ArgumentNullException(nameof(permission));
+
+                var leftFilter = this.LeftBuilder.GetSecurityFilterExpression(permission);
+                var rightFilter = this.RightBuilder.GetSecurityFilterExpression(permission);
+
+                return this.BuildOperation(leftFilter, rightFilter);
+            }
 
             public override Expression<Func<IPermission<TIdent>, bool>> GetAccessorsFilter(TDomainObject domainObject, HierarchicalExpandType expandType)
             {
@@ -523,16 +536,6 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             }
 
 
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
-            {
-                var leftFilter = this.LeftBuilder.GetSecurityFilterExpression(expandType).ExpandConst().InlineEval();
-                var rightFilter = this.RightBuilder.GetSecurityFilterExpression(expandType).ExpandConst().InlineEval();
-
-                return (domainObject, permission) =>
-
-                               leftFilter.Eval(domainObject, permission) && rightFilter.Eval(domainObject, permission);
-            }
-
             protected override Expression<Func<TArg, bool>> BuildOperation<TArg>(Expression<Func<TArg, bool>> arg1, Expression<Func<TArg, bool>> arg2)
             {
                 return arg1.BuildAnd(arg2);
@@ -546,15 +549,7 @@ namespace Framework.SecuritySystem.Rules.Builders.V2
             {
 
             }
-            public override Expression<Func<TDomainObject, IPermission<TIdent>, bool>> GetSecurityFilterExpression(HierarchicalExpandType expandType)
-            {
-                var leftFilter = this.LeftBuilder.GetSecurityFilterExpression(expandType).ExpandConst().InlineEval();
-                var rightFilter = this.RightBuilder.GetSecurityFilterExpression(expandType).ExpandConst().InlineEval();
 
-                return (domainObject, permission) =>
-
-                               leftFilter.Eval(domainObject, permission) || rightFilter.Eval(domainObject, permission);
-            }
 
             protected override Expression<Func<TArg, bool>> BuildOperation<TArg>(Expression<Func<TArg, bool>> arg1, Expression<Func<TArg, bool>> arg2)
             {
