+++ src/Framework.SecuritySystem/SecurityPath/Builders/V1_MaterializedPermissions/SecurityExpressionFilter.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Linq.Expressions;
 
@@ -8,7 +7,7 @@ using Framework.Core;
 using Framework.Persistent;
 
 
-namespace Framework.SecuritySystem.Rules.Builders.V3
+namespace Framework.SecuritySystem.Rules.Builders.MaterializedPermissions
 {
     public class SecurityExpressionFilter<TPersistentDomainObjectBase, TDomainObject, TSecurityOperationCode, TIdent> : ISecurityExpressionFilter<TDomainObject>
 
@@ -20,19 +19,20 @@ namespace Framework.SecuritySystem.Rules.Builders.V3
 
         private readonly Lazy<Func<TDomainObject, IEnumerable<string>>> getAccessorsFunc;
 
-        // ReSharper disable once StaticMemberInGenericType
-        [SuppressMessage("SonarQube", "S2743")]
-        private static readonly ILambdaCompileCache LambdaCompileCache = new LambdaCompileCache();
+        private static readonly ILambdaCompileCache LambdaCompileCache = new LambdaCompileCache(LambdaCompileMode.All);
 
         public SecurityExpressionFilter(
             SecurityExpressionBuilderBase<TPersistentDomainObjectBase, TDomainObject, TIdent> builder,
             ContextSecurityOperation<TSecurityOperationCode> securityOperation)
         {
             if (builder == null) throw new ArgumentNullException(nameof(builder));
-            if (securityOperation == null) throw new
-                    ArgumentNullException(nameof(securityOperation));
+            if (securityOperation == null) throw new ArgumentNullException(nameof(securityOperation));
 
-            var filterExpression = builder.GetSecurityFilterExpression(securityOperation).ExpandConst().InlineEval();
+            var usedTypes = builder.GetUsedTypes().Distinct();
+
+            var permissions = builder.Factory.AuthorizationSystem.GetPermissions(securityOperation, usedTypes);
+
+            var filterExpression = builder.GetSecurityFilterExpression(permissions);
 
             this.InjectFunc = q => q.Where(filterExpression);
 
@@ -56,7 +56,6 @@ namespace Framework.SecuritySystem.Rules.Builders.V3
         public Func<IQueryable<TDomainObject>, IQueryable<TDomainObject>> InjectFunc { get; }
 
         public Func<TDomainObject, bool> HasAccessFunc { get; }
-
 
         public IEnumerable<string> GetAccessors(TDomainObject domainObject)
         {
