+++ src/Framework.SecuritySystem/SecurityPath/Builders/V2_QueryablePermissions/SecurityExpressionBuilderFactory.cs
@@ -5,7 +5,7 @@ using Framework.Persistent;
 
 using JetBrains.Annotations;
 
-namespace Framework.SecuritySystem.Rules.Builders.V3
+namespace Framework.SecuritySystem.Rules.Builders.QueryablePermissions
 {
     public class SecurityExpressionBuilderFactory<TPersistentDomainObjectBase, TIdent> : SecurityExpressionBuilderFactoryBase<TPersistentDomainObjectBase, TIdent>
